+++ src/wavegame/FastEnemy.java
@@ -17,7 +17,7 @@ import java.util.Random;
 public class FastEnemy extends GameObject {
     private Handler handler;
     
-    public FastEnemy(int x, int y, ID id, Handler handler){
+    public FastEnemy(float x, float y, ID id, Handler handler){
         super(x, y, id);
         this.handler= handler;
         
@@ -28,7 +28,7 @@ public class FastEnemy extends GameObject {
     }
     
     public Rectangle getBounds(){
-        return new Rectangle(x, y, 16, 16);
+        return new Rectangle((int)x, (int)y, 16, 16);
     }
     
     public void tick(){
@@ -43,6 +43,6 @@ public class FastEnemy extends GameObject {
     
     public void render(Graphics g){
         g.setColor(Color.red);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int)x, (int)y, 16, 16);
     }
 }
